+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/tests/latest/test_storage_batch_operations.py
@@ -5,7 +5,7 @@
 
 import os
 from azure.cli.testsdk import LiveScenarioTest, StorageAccountPreparer, ResourceGroupPreparer, JMESPathCheck
-from .storage_test_util import StorageScenarioMixin, StorageTestFilesPreparer
+from ..storage_test_util import StorageScenarioMixin, StorageTestFilesPreparer
 
 
 class StorageBatchOperationScenarios(StorageScenarioMixin, LiveScenarioTest):