+++ src/command_modules/azure-cli-keyvault/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+2.2.3
++++++
+* Fixed Key Vault key import for RSA keys
+
 2.2.2
 +++++
 * adding commands for managing storage accounts and sas-definitions